Update #2
In this update I've created a new config file for secure the stream with
username and password!
To use it you need to create a htpasswd file with the credentials for as
many users you want.
I've also included the script for runinng the http server and a .rules
file to give permissions to the camera.